% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLcontainer.R
\docType{class}
\name{Areskuel-class}
\alias{Areskuel-class}
\alias{Areskuel}
\title{An S4 class to represent the SQL query return value}
\description{
An S4 class to represent the SQL query return value
}
\section{Slots}{

\describe{
\item{\code{server}}{The server connnected to at the time of the query execution.}

\item{\code{database}}{The database connected to at the time of the query execution.}

\item{\code{sql}}{The SQL query.}

\item{\code{data}}{The resulting data.}
}}

