% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{areskuel-package}
\alias{areskuel}
\alias{areskuel-package}
\title{areskuel: Interface to Podman Containers Running Microsoft SQL Server}
\description{
Easily set up, manage and interact with podman containers that run Microsoft SQL Server. RSQL was already taken on CRAN...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dheimgartner/areskuel}
  \item Report bugs at \url{https://github.com/dheimgartner/areskuel/issues}
}

}
\author{
\strong{Maintainer}: Daniel Heimgartner \email{d.heimgartners@gmail.com} [copyright holder]

}
