% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sys_dependencies.R
\name{check_sys_dependencies}
\alias{check_sys_dependencies}
\title{Checks if system dependencies ara available}
\usage{
check_sys_dependencies()
}
\value{
Named list containing boolean values
}
\description{
Checks if system dependencies ara available
}
