% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{RSQL-package}
\alias{RSQL-package}
\title{RSQL: Interface to Podman Containers Running Microsoft SQL Server}
\description{
Easily set up, manage and interact with podman containers that run Microsoft SQL Server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dheimgartner/RSQL}
  \item Report bugs at \url{https://github.com/dheimgartner/RSQL/issues}
}

}
\author{
\strong{Maintainer}: Daniel Heimgartner \email{d.heimgartners@gmail.com}

}
