% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLcontainer.R
\name{SQLcontainer$set_database}
\alias{SQLcontainer$set_database}
\title{Set the MSSQL database to connect to}
\arguments{
\item{database}{The database to connect to.}
}
\description{
Set the MSSQL database to connect to
}
