% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLcontainer.R
\name{SQLcontainer$set_server}
\alias{SQLcontainer$set_server}
\title{Set the MSSQL server to connect to}
\arguments{
\item{server}{The server to connect to.}
}
\description{
Set the MSSQL server to connect to
}
