% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLcontainer.R
\docType{class}
\name{SQLcontainer}
\alias{SQLcontainer}
\title{The main reference class to manage and interact with the podman container}
\description{
...and in paricular it's SQL server!
}
\section{Fields}{

\describe{
\item{\code{path}}{The path to the \code{mssqlcontainer.sh} executable.}

\item{\code{container}}{The name of the container. Defaults to \code{"areskuel"}.}

\item{\code{user}}{The SQL user. Defaults to \code{"uaresql"}.}

\item{\code{con}}{The DB connection.}

\item{\code{server}}{The server to connect to. Defaults to \code{"localhost"}.}

\item{\code{database}}{The database to connect to. Defaults to \code{"master"}.}

\item{\code{port}}{The port number. Defaults to \code{"1433"}.}

\item{\code{sql_history}}{A list of \code{"SQLhistory"} elements \link{SQLhistory}.}

\item{\code{password}}{The associated password for the user. Defaults to \code{"P@assword123!"}.
Needs to conform with MSSQL's password policy!}

\item{\code{driver}}{The \code{odbc}-driver to use. Defaults to \code{"ODBC Driver 18 for SQL Server"}.}
}}


\examples{
\donttest{

## check sys prerequisites
checks <- check_sys_dependencies()
if (!all(checks)) {
  stop("System requirements not met. Make sure podman and ODBC drivers are installed.")
}

container <- SQLcontainer(
  container = "test_areskuel",
  user = "test_user"
)

container$create()
container$connect()
container$con

## create a test table "iris" with the iris data
container$test()

rsql <- container$sql("SELECT * FROM iris")
rsql

## create a new database
foo <- container$sql("CREATE DATABASE db_foo")
container$set_database("db_foo")
container$con

## check the history
container$sql_history

## close the connection
container$disconnect()

## delete the container (all data will be lost!)
container$delete()

}
}
\seealso{
The methods documentation and \code{SQLcontainer$help()}
}
