% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{RSQL-package}
\alias{RSQL-package}
\title{RSQL: Interface to a podman container running Microsoft SQL Server}
\description{
Allows R users to set up a podman container running Microsoft SQL Server and interact with it.
}
\author{
\strong{Maintainer}: Daniel Heimgartner \email{d.heimgartners@gmail.com}

}
